+++ ios/Widget/Widget.swift
@@ -2,20 +2,6 @@ import WidgetKit
 import SwiftUI
 import Intents
 
-fileprivate var calendar: Calendar {
-  var calendar = Calendar(identifier: .gregorian)
-  calendar.locale = .autoupdatingCurrent
-  calendar.timeZone = .autoupdatingCurrent
-  return calendar
-}
-
-fileprivate var dateFormater: DateFormatter {
-  let dateFormater = DateFormatter()
-  dateFormater.locale = .autoupdatingCurrent
-  dateFormater.timeZone = .autoupdatingCurrent
-  return dateFormater
-}
-
 struct Provider: TimelineProvider {
   func placeholder(in context: Context) -> Entry {
     .init(date: .now)
@@ -35,100 +21,6 @@ struct Provider: TimelineProvider {
   }
 }
 
-struct Entry: TimelineEntry {
-  // Timeline Entry required property
-  let date: Date
-
-  // PillSheet property
-  let pillSheetTodayPillNumber: Int?
-  let pillSheetGroupTodayPillNumber: Int?
-  let pillSheetEndDisplayPillNumber: Int?
-  let pillSheetLastTakenDate: Date?
-
-  // Setting property
-  var settingPillSheetAppearanceMode: SettingPillSheetAppearanceMode = .number
-  enum SettingPillSheetAppearanceMode: String {
-    case number, date, sequential
-  }
-
-  // Timestamp
-  let pillSheetValueLastUpdateDateTime: Date?
-
-  init(date: Date) {
-    self.date = date
-
-    func contains(_ key: String) -> Bool {
-      UserDefaults(suiteName: Plist.appGroupKey)?.dictionaryRepresentation().keys.contains(key) == true
-    }
-
-    if contains(Const.pillSheetTodayPillNumber), let pillSheetTodayPillNumber = UserDefaults(suiteName: Plist.appGroupKey)?.integer(forKey: Const.pillSheetTodayPillNumber) {
-      self.pillSheetTodayPillNumber = pillSheetTodayPillNumber
-    } else {
-      self.pillSheetTodayPillNumber = nil
-    }
-
-    if contains(Const.pillSheetGroupTodayPillNumber), let pillSheetGroupTodayPillNumber = UserDefaults(suiteName: Plist.appGroupKey)?.integer(forKey: Const.pillSheetGroupTodayPillNumber) {
-      self.pillSheetGroupTodayPillNumber = pillSheetGroupTodayPillNumber
-    } else {
-      self.pillSheetGroupTodayPillNumber = nil
-    }
-
-    if contains(Const.pillSheetEndDisplayPillNumber), let pillSheetEndDisplayPillNumber = UserDefaults(suiteName: Plist.appGroupKey)?.integer(forKey: Const.pillSheetEndDisplayPillNumber) {
-      self.pillSheetEndDisplayPillNumber = pillSheetEndDisplayPillNumber
-    } else {
-      self.pillSheetEndDisplayPillNumber = nil
-    }
-
-    if contains(Const.pillSheetLastTakenDate), let pillSheetLastTakenDateEpochMilliSecond = UserDefaults(suiteName: Plist.appGroupKey)?.integer(forKey: Const.pillSheetLastTakenDate) {
-      self.pillSheetLastTakenDate = Date(timeIntervalSince1970: TimeInterval(pillSheetLastTakenDateEpochMilliSecond / 1000))
-    } else {
-      self.pillSheetLastTakenDate = nil
-    }
-
-    if contains(Const.settingPillSheetAppearanceMode), let settingPillSheetAppearanceMode = UserDefaults(suiteName: Plist.appGroupKey)?.string(forKey: Const.settingPillSheetAppearanceMode) {
-      self.settingPillSheetAppearanceMode = .init(rawValue: settingPillSheetAppearanceMode) ?? .number
-    }
-
-    if contains(Const.pillSheetValueLastUpdateDateTime), let pillSheetValueLastUpdateDateTimeEpochMilliSecond = UserDefaults(suiteName: Plist.appGroupKey)?.integer(forKey: Const.pillSheetValueLastUpdateDateTime) {
-      self.pillSheetValueLastUpdateDateTime = Date(timeIntervalSince1970: TimeInterval(pillSheetValueLastUpdateDateTimeEpochMilliSecond / 1000))
-    } else {
-      self.pillSheetValueLastUpdateDateTime = nil
-    }
-  }
-
-  var todayPillNumber: Int? {
-    guard let pillSheetValueLastUpdateDateTime = pillSheetValueLastUpdateDateTime else {
-      return nil
-    }
-
-    let todayPillNumberBase: Int
-    switch settingPillSheetAppearanceMode {
-    case .number, .date:
-      guard let pillSheetTodayPillNumber = pillSheetTodayPillNumber else {
-        return nil
-      }
-      todayPillNumberBase = pillSheetTodayPillNumber
-    case .sequential:
-      guard let recordedPillSheetGroupTodayPillNumber = pillSheetGroupTodayPillNumber else {
-        return nil
-      }
-      todayPillNumberBase = recordedPillSheetGroupTodayPillNumber
-    }
-
-    guard let diff = calendar.dateComponents([.day], from: date, to: pillSheetValueLastUpdateDateTime).day else {
-      return todayPillNumberBase
-    }
-    let todayPillNumber = todayPillNumberBase + diff
-
-    if let pillSheetEndDisplayPillNumber = pillSheetEndDisplayPillNumber, todayPillNumber > pillSheetEndDisplayPillNumber {
-      // 更新されるまで常に1で良い
-      return 1
-    } else {
-      return todayPillNumber
-    }
-  }
-
-}
 
 struct WidgetEntryView : View {
   var entry: Provider.Entry
